PC-41 moved user specific functions to new user file in preparation for project resturcture
